implement action limitations

+++ src/main/java/com/alrex/parcool/common/action/StaminaConsumption.java
@@ -1,4 +1,4 @@
 package com.alrex.parcool.common.action;
 
-public record StaminaConsumption(int onStart, int onWorking, int onFinish) {
+public record StaminaConsumption(short onStart, short onWorking, short onFinish) {
 }
